+++ src/index.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import { BrowserRouter, Route } from 'react-router-dom';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 import './sass/style.css';
 
 // Import components
-import App from './components/App/App';
-import Sidebar from './components/Sidebar/Sidebar';
-import Auth from './components/Auth/Auth';
-import ClientList from './components/ClientList/ClientList';
-import NewClientForm from './components/NewClientForm/NewClientForm';
+import App from './components/App';
+import Sidebar from './components/Sidebar';
+import Auth from './components/Auth';
+import ClientList from './components/ClientList';
+import NewClientForm from './components/NewClientForm';
 
 import store from './store';
 
@@ -18,14 +18,16 @@ const router = (
     <Provider store={store}>
         <App>
             <BrowserRouter>
-                <div>
+                <div className="app">
                     <Sidebar>
                         <Auth />
                         <ClientList />
                     </Sidebar>
 
-                    {/* <Route exact path="/" component={() => <h1>Home</h1>)} /> */}
-                    {/* <Route path="" */}
+                    <Switch className="appmain">
+                        <Route exact path="/" component={() => <h1>Home</h1>} />
+                        <Route path="/client/:clientId" component={() => <h1>Clienbt</h1>} />
+                    </Switch>
 
                     <NewClientForm />
                 </div>
