+++ src/sass/_layout.scss
@@ -1,5 +1,5 @@
-@import '../../sass/variables';
-@import '../../sass/mixins';
+@import './variables';
+@import './mixins';
 
 .app {
     display: flex;
