+++ src/sass/style.scss
@@ -4,3 +4,4 @@
 @import 'base';
 @import 'type';
 @import 'forms';
+@import 'layout';
